+++ EnvironmentSimulator/Applications/replayer/Replay.cpp
@@ -429,7 +429,7 @@ int Replay::ParsePackets()
                 }
                 case static_cast<id_t>(Dat::PacketId::SPEED):
                 {
-                    float speed;
+                    double speed;
                     if (dat_reader_->ReadPacket(gp, speed) != 0)
                     {
                         LOG_ERROR("Failed reading speed data.");
@@ -500,7 +500,7 @@ int Replay::ParsePackets()
                 }
                 case static_cast<id_t>(Dat::PacketId::WHEEL_ANGLE):
                 {
-                    float wheel_angle;
+                    double wheel_angle;
                     if (dat_reader_->ReadPacket(gp, wheel_angle) != 0)
                     {
                         LOG_ERROR("Failed reading wheel angle.");
@@ -511,7 +511,7 @@ int Replay::ParsePackets()
                 }
                 case static_cast<id_t>(Dat::PacketId::WHEEL_ROT):
                 {
-                    float wheel_rot;
+                    double wheel_rot;
                     if (dat_reader_->ReadPacket(gp, wheel_rot) != 0)
                     {
                         LOG_ERROR("Failed reading wheel rotation.");
@@ -594,7 +594,7 @@ int Replay::ParsePackets()
                 }
                 case static_cast<id_t>(Dat::PacketId::POS_OFFSET):
                 {
-                    float offset;
+                    double offset;
                     if (dat_reader_->ReadPacket(gp, offset) != 0)
                     {
                         LOG_ERROR("Failed reading position offset.");
@@ -605,7 +605,7 @@ int Replay::ParsePackets()
                 }
                 case static_cast<id_t>(Dat::PacketId::POS_T):
                 {
-                    float t;
+                    double t;
                     if (dat_reader_->ReadPacket(gp, t) != 0)
                     {
                         LOG_ERROR("Failed reading position T.");
@@ -616,7 +616,7 @@ int Replay::ParsePackets()
                 }
                 case static_cast<id_t>(Dat::PacketId::POS_S):
                 {
-                    float s;
+                    double s;
                     if (dat_reader_->ReadPacket(gp, s) != 0)
                     {
                         LOG_ERROR("Failed reading position S.");
@@ -643,7 +643,7 @@ int Replay::ParsePackets()
                 }
                 case static_cast<id_t>(Dat::PacketId::REFPOINT_X_OFFSET):
                 {
-                    float refpoint_x_offset;
+                    double refpoint_x_offset;
                     if (dat_reader_->ReadPacket(gp, refpoint_x_offset) != 0)
                     {
                         LOG_ERROR("Failed reading refpoint_x_offset");
@@ -654,7 +654,7 @@ int Replay::ParsePackets()
                 }
                 case static_cast<id_t>(Dat::PacketId::MODEL_X_OFFSET):
                 {
-                    float model_x_offset;
+                    double model_x_offset;
                     if (dat_reader_->ReadPacket(gp, model_x_offset) != 0)
                     {
                         LOG_ERROR("Failed reading model_x_offset");
@@ -1300,7 +1300,7 @@ ReplayEntry Replay::GetReplayEntryAtTimeIncremental(int id, double t) const
     const auto& timeline = objects_timeline_.at(id);
 
     entry.state.info.id             = id;
-    entry.state.info.timeStamp      = static_cast<float>(t);
+    entry.state.info.timeStamp      = t;
     entry.state.info.model_id       = timeline.model_id_.get_value_incremental(t).value_or(-1);
     entry.state.info.obj_type       = timeline.obj_type_.get_value_incremental(t).value();
     entry.state.info.obj_category   = timeline.obj_category_.get_value_incremental(t).value();
@@ -1336,7 +1336,7 @@ ReplayEntry Replay::GetReplayEntryAtTimeBinary(int id, double t) const
     const auto& timeline = objects_timeline_.at(id);
 
     entry.state.info.id             = id;
-    entry.state.info.timeStamp      = static_cast<float>(t);
+    entry.state.info.timeStamp      = t;
     entry.state.info.model_id       = timeline.model_id_.get_value_binary(t).value_or(-1);
     entry.state.info.obj_type       = timeline.obj_type_.get_value_binary(t).value();
     entry.state.info.obj_category   = timeline.obj_category_.get_value_binary(t).value();
