+++ EnvironmentSimulator/Applications/replayer/Replay.hpp
@@ -301,22 +301,22 @@ namespace scenarioengine
         Timeline<int>                     obj_category_;
         Timeline<int>                     ctrl_type_;
         Timeline<std::string>             name_;
-        Timeline<float>                   speed_;
-        Timeline<float>                   wheel_angle_;
-        Timeline<float>                   wheel_rot_;
+        Timeline<double>                  speed_;
+        Timeline<double>                  wheel_angle_;
+        Timeline<double>                  wheel_rot_;
         Timeline<OSCBoundingBox>          bounding_box_;
         Timeline<int>                     scale_mode_;
         Timeline<int>                     visibility_mask_;
         Timeline<Dat::Pose>               pose_;
         Timeline<id_t>                    road_id_;
         Timeline<int>                     lane_id_;
-        Timeline<float>                   pos_offset_;
-        Timeline<float>                   pos_t_;
-        Timeline<float>                   pos_s_;
+        Timeline<double>                  pos_offset_;
+        Timeline<double>                  pos_t_;
+        Timeline<double>                  pos_s_;
         Timeline<bool>                    active_;
-        Timeline<float>                   odometer_;
-        Timeline<float>                   refpoint_x_offset_;
-        Timeline<float>                   model_x_offset_;
+        Timeline<double>                  odometer_;
+        Timeline<double>                  refpoint_x_offset_;
+        Timeline<double>                  model_x_offset_;
         Timeline<std::string>             model3d_;
         Timeline<std::vector<SE_Point2D>> outline_;
     };
@@ -344,11 +344,11 @@ namespace scenarioengine
         int            obj_type;      // 0=None, 1=Vehicle, 2=Pedestrian, 3=MiscObj (see Object::Type enum)
         int            obj_category;  // sub type for vehicle, pedestrian and miscobj
         int            ctrl_type;     // See Controller::Type enum
-        float          timeStamp;
+        double         timeStamp;
         std::string    name;
-        float          speed;
-        float          wheel_angle;  // Only used for vehicle
-        float          wheel_rot;    // Only used for vehicle
+        double         speed;
+        double         wheel_angle;  // Only used for vehicle
+        double         wheel_rot;    // Only used for vehicle
         OSCBoundingBox boundingbox;
         int            scaleMode;       // 0=None, 1=BoundingBoxToModel, 2=ModelToBoundingBox (see enum EntityScaleMode)
         int            visibilityMask;  // bitmask according to Object::Visibility (1 = Graphics, 2 = Traffic, 4 = Sensors)
